buffer: Rename buffer_init to buffer_init_stream
Name of the function was changed because it is only responsible for
initiating a stream. The caps parameter was omitted because it was only
assigned to a structure field. This field is now assigned by the function
caller. In the second call of the function, the current value of the field
was passed as a parameter, which resulted in the unnecessary assignment
of the same value to the structure field.

Signed-off-by: Adrian Warecki <adrian.warecki@intel.com>