chain_dma: xrun notificaton update
Change xrun notifications to use the notification pool. Improve the logic
for clearing the flag that informs the notification was sent. Previously,
it was cleared when xrun occurred and the notification was sent,
which meant that a notification was sent every second time xrun was
detected, which could cause a flood of notifications.

Signed-off-by: Adrian Warecki <adrian.warecki@intel.com>