ipc4: Add notification pool
Add a pool of IPC messages for sending notifications. These messages are
dynamically allocated as needed. Add a callback to the ipc_msg structure
that is called after sending a message, allowing messages to be returned
to the pool once sent.

Using a notification pool simplifies the process of sending sporadic
notifications by components. This approach eliminates the need for
components to manage the allocation and release of notifications.
Additionally, it conserves memory, as notifications that are never sent
during normal operation will not be allocated.

Signed-off-by: Adrian Warecki <adrian.warecki@intel.com>